+++ Sources/TuistSupport/Constants.swift
@@ -26,6 +26,11 @@ public enum Constants {
     public static let joinSlackURL: String = "https://slack.tuist.io/"
     public static let tuistGeneratedFileName = ".tuist-generated"
 
+    /// The cache version.
+    /// This should change only when it changes the logic to map a `TuistGraph.Target` to a cached build artifact.
+    /// Changing this results in changing the target hash and hence forcing a rebuild of its artifact.
+    public static let cacheVersion = "1.0.0"
+
     public enum Vendor {
         public static let swiftLint = "swiftlint"
         public static let swiftDoc = "swift-doc"